fabric_notebooks/requirements.txt
# Python Package Requirements for G4S API Sync

# Core data processing
pyspark>=3.3.0
delta-spark>=2.2.0

# API and HTTP
requests>=2.28.0
urllib3>=1.26.0

# Azure integrations
azure-identity>=1.12.0
azure-keyvault-secrets>=4.6.0

# Data serialization
pyarrow>=10.0.0

# Note: In Microsoft Fabric, many of these packages are pre-installed
# This file is provided for reference and local development only
